Align A2DP and speaker volume in tandem case
Absolute volume is used for supported A2DP devices.
This forces the specific stream volume to always follow
a specified volume curve. This is fine for volumes that
only applies to the A2DP output device but not for tandem
usecases. When toggle from volume index 0 to 1, first volume
index for speaker was applied and then overriden by A2DP volume.

When A2DP and speaker are used in tandem, volume should follow
speaker.

Test: Pair BT headset, toggle ringtone volume between 0 and 1 and test
Bug: 293982417
Change-Id: I854d2de1abd200778fdf5a3af3aa90db55d2b71e